chore: move old NewInUnleash to legacy file
